+++ sdk/cosmos/azure_data_cosmos/CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Added basic multi-region writes support. ([#3482](https://github.com/Azure/azure-sdk-for-rust/pull/3482) and [#3495](https://github.com/Azure/azure-sdk-for-rust/pull/3495))
 - Added new `CosmosResponse` that wraps `azure_core::Response` for all operations except queries. ([#3622](https://github.com/Azure/azure-sdk-for-rust/pull/3622))
+- Added transactional batch support for executing multiple operations atomically within the same partition key. ([#XXXX](https://github.com/Azure/azure-sdk-for-rust/pull/XXXX))
 
 ### Breaking Changes